refactor!: hoist metadatumToCip25 to NftMetadata.fromMetadatum
NftMetadata type shouldn't be coupled to cip25, because there are
now 2 sources for NFT metadata:
- cip25: auxiliary data (tx metadata)
- cip68/222: datum
