+++ packages/core/src/Asset/NftMetadata/types.ts
