+++ packages/core/src/Asset/index.ts
@@ -1,3 +1,4 @@
 export * as util from './util';
 export * from './types';
 export * from './cip67';
+export * from './NftMetadata';
