+++ packages/core/src/Asset/types/index.ts
@@ -1,3 +1,2 @@
-export * from './NftMetadata';
 export * from './TokenMetadata';
 export * from './AssetInfo';
