+++ packages/core/src/Asset/util/index.ts
@@ -1,4 +1,3 @@
-export * from './metadatumToCip25';
 export * from './coalesceTokenMaps';
 export * from './removeNegativesFromTokenMap';
 export * from './subtractTokenMaps';
